requirements.txt
azure_storage==0.37.0
detectron2==0.6+cpu
fastapi==0.96.1
loguru==0.7.0
numpy==1.20.3
opencv_contrib_python==4.4.0.42
pandas==1.3.5
Pillow==8.3.2
Pillow==9.5.0
pydantic==1.10.8
PyYAML==5.4.1
SQLAlchemy==1.4.46
streamlit==0.89.0
torch==1.9.0
uvicorn==0.22.0
